Add files via upload
Updated Slides as of September 18, 2019
Added: Better views of the game and introducing the Loading Screen implementation.
Coming Up: An actual DEMO of the VR navigation and loading in action with the world being designed in the background.
Coming Coming Up: The First Puzzle's implementation. The puzzles will take all October, leaving November to add finishing touches as well a flavor.